% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp21032.R
\name{knapsack_problem}
\alias{knapsack_problem}
\title{best choice of bag}
\usage{
knapsack_problem(n, c, w, v)
}
\arguments{
\item{n}{the sum of stuff}

\item{c}{max weight of the bag}

\item{w}{the weight of each stuff}

\item{v}{the value of each stuff}
}
\description{
best choice of bag
}
\examples{
\dontrun{
n = 6
c = 10
w = c(2, 1, 1, 5, 4, 3)
v = c(2, 2, 3, 1, 5, 2)
knapsack_problem(n,c,w,v)
}
}
