% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp21032.R
\name{prob_point}
\alias{prob_point}
\title{probability of specific point}
\usage{
prob_point(n, k)
}
\arguments{
\item{n}{specific point}

\item{k}{the number of prob}
}
\description{
probability of specific point
}
\examples{
\dontrun{
n = 10
k = 2
prob_point(n,k)
}
}
