\name{StatComp21032-package}
\alias{StatComp21032-package}
\alias{StatComp21032}
\docType{package}
\title{
  compute the probability of specific point and best choice of bag
}
\description{
  compute the probability of specific point and best choice of bag
}
\details{
  compute the probability of specific point and best choice of bag
}
\author{
Mengge Yang

Maintainer: Mengge Yang <yangmg@mail.ustc.edu.cn>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
